Init chunk system.

+++ VM/vm.h
@@ -1,13 +1,13 @@
 #ifndef GOBLIN_VM_H
 #define GOBLIN_VM_H
 
-#include "value.h"
+#include "chunk.h"
 
 #define STACK_MAX 256
 
 typedef struct {
   Chunk chunk;
-  int ip;
+  uint8_t ip;
   Value stack[STACK_MAX];
   Value* stackTop;
 } VM;
